+++ backend/canisters/community/api/src/updates/c2c_bot_send_message.rs
@@ -1,6 +1,8 @@
 use crate::send_message;
 use serde::{Deserialize, Serialize};
-use types::{BotInitiator, BotMessageContent, ChannelId, EventIndex, GroupReplyContext, MessageId, MessageIndex, UserId};
+use types::{
+    BotInitiator, BotMessageContent, ChannelId, EventIndex, GroupReplyContext, MessageId, MessageIndex, OgPreview, UserId,
+};
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -14,6 +16,8 @@ pub struct Args {
     pub bot_name: String,
     pub block_level_markdown: bool,
     pub finalised: bool,
+    #[serde(default)]
+    pub og_previews: Vec<OgPreview>,
 }
 
 impl From<Args> for send_message::Args {
@@ -33,6 +37,7 @@ impl From<Args> for send_message::Args {
             channel_rules_accepted: None,
             message_filter_failed: None,
             new_achievement: false,
+            og_previews: value.og_previews,
         }
     }
 }