+++ backend/canisters/group/impl/src/updates/send_message.rs
@@ -103,6 +103,7 @@ pub(crate) fn send_message_impl(
             queue: &mut state.data.local_user_index_event_sync_queue,
         },
         finalised,
+        args.og_previews,
         now,
     )?;
 
@@ -150,6 +151,7 @@ fn c2c_send_message_impl(args: C2CArgs, state: &mut RuntimeState) -> OCResult<Su
             queue: &mut state.data.local_user_index_event_sync_queue,
         },
         true,
+        args.og_previews,
         now,
     )?;