+++ backend/canisters/user/api/src/updates/edit_message_v2.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
-use types::{MessageContentInitial, MessageId, MessageIndex, UnitResult, UserId};
+use types::{MessageContentInitial, MessageId, MessageIndex, OgPreview, UnitResult, UserId};
 
 #[ts_export(user, edit_message)]
 #[derive(Serialize, Deserialize, Debug)]
@@ -10,6 +10,8 @@ pub struct Args {
     pub message_id: MessageId,
     pub content: MessageContentInitial,
     pub block_level_markdown: Option<bool>,
+    #[serde(default)]
+    pub og_previews: Vec<OgPreview>,
 }
 
 pub type Response = UnitResult;