+++ backend/canisters/user/impl/src/updates/edit_message.rs
@@ -31,6 +31,7 @@ fn edit_message_impl(args: Args, state: &mut RuntimeState) -> OCResult {
             message_id: args.message_id,
             content: args.content.clone().into(),
             block_level_markdown: args.block_level_markdown,
+            og_previews: args.og_previews.clone(),
             finalise_bot_message: false,
             now,
         };
@@ -54,6 +55,7 @@ fn edit_message_impl(args: Args, state: &mut RuntimeState) -> OCResult {
                     message_id: args.message_id,
                     content: args.content.into(),
                     block_level_markdown: args.block_level_markdown,
+                    og_previews: args.og_previews,
                 })),
             );