+++ backend/canisters/user_index/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Add `og_previews` field to messages for handling OpenGraph previews ([#9002](https://github.com/open-chat-labs/open-chat/pull/9002))
 - Add `set_openai_api_key` endpoint ([#9021](https://github.com/open-chat-labs/open-chat/pull/9021))
 
 ### Changed